feat : include fastapi router and add auth dependency to exisiting routes

+++ backend/src/app.py
@@ -6,14 +6,15 @@ import uuid
 from fastapi import FastAPI,HTTPException,File,Form,UploadFile,Depends
 import imagekitio
 
-from src.schemas import PostFormat
-from src.db import Post,create_db_and_tables,get_async_session
+from src.schemas import PostFormat,UserRead,UserCreate,UserUpdate
+from src.db import Post,create_db_and_tables,get_async_session,User
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
 from contextlib import asynccontextmanager
 
 from src.images import imagekit
-# from imagekitio.models.UploadFileRequestOptions import UploadFileRequestOptions
+#!from imagekitio.models.UploadFileRequestOptions import UploadFileRequestOptions >>deprecated
+from src.users import auth_backend,current_active_user,fastapi_users
 
 
 @asynccontextmanager
@@ -22,6 +23,12 @@ async def lifespan(app:FastAPI):
     yield #wtf is yield?
 
 app=FastAPI(lifespan=lifespan)
+
+app.include_router(fastapi_users.get_auth_router(auth_backend),prefix="/auth/jwt",tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead,UserCreate),prefix="/auth",tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(),prefix="/auth",tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(UserRead),prefix="/auth",tags=["auth"])
+app.include_router(fastapi_users.get_users_router(UserRead,UserUpdate),prefix="/users",tags=["users"])
 
 posts : dict = {
     1: {
@@ -75,20 +82,29 @@ posts : dict = {
 
 @app.get("/feed")
 async def get_posts(
-    session:AsyncSession=Depends(get_async_session)
+    session:AsyncSession=Depends(get_async_session),
+    user:User=Depends(current_active_user),
 ):
     results=await session.execute(select(Post).order_by(Post.created_at.desc()))
 
     posts=[row[0] for row in results.all()] #?why do we access row[0]
     posts_data=[]
+
+    user_data=await session.execute(select(User))
+    users=[row[0] for row in user_data.all()]
+    user_dict={user.id:user.email for user in users}
+
     for post in posts:
         posts_data.append({
             "id":str(post.id),
+            "user":str(post.user_id),
             "caption":post.caption,
             "url":post.url,
             "file_type":post.file_type,
             "file_name":post.file_name,
-            "created_at":post.created_at.isoformat()
+            "created_at":post.created_at.isoformat(),
+            "is_owner":post.user_id==user.id,
+            "email":user_dict.get(post.user_id,"Unknown User")
         })
 
     return {"posts":posts_data}
@@ -99,6 +115,7 @@ async def get_posts(
 async def upload_file(
     file:UploadFile=File(...),
     caption:str=Form(""),
+    user:User=Depends(current_active_user),
     session:AsyncSession=Depends(get_async_session)
 ):
     temp_file_path=None
@@ -118,6 +135,7 @@ async def upload_file(
 
 
         post=Post(
+            user_id=user.id,
             caption=caption,
             url=upload_result.url,
             file_type="video" if file.content_type.startswith("video/") else "image",
@@ -142,6 +160,7 @@ async def upload_file(
 @app.delete("/posts/{post_id}")
 async def delete_post(
     post_id:str,
+    user:User=Depends(current_active_user),
     session:AsyncSession=Depends(get_async_session)
     ):
     try:
@@ -153,6 +172,9 @@ async def delete_post(
 
         if not post:
             raise HTTPException(status_code=404,detail="post not found")
+        
+        if post.user_id!=user.id:
+            raise HTTPException(status_code=403,detail="you don't have permission to delete this post")
         
         await session.delete(post)
         await session.commit()
